SDK: Refactor out DDSTask
